refactor(path): Kakao API의 주소 정보 클래스를 DTO로 이동 - API 규격을 맞추기 위한 클래스이므로 VO로 정의하기 부적합하다고 판단됨
